test(errors): consolidate legacy retry regression
